requirements.txt
streamlit_lottie
streamlit
streamlit-option-menu
requests
Pillow